pull signatures into their own file. remove stale exporters code to create signatures. refactor elmo exporter to return new expected output
